simple 3d cube render

+++ led_panel/Core/Inc/renderer.h
@@ -44,7 +44,7 @@ void set_pixel_w(int x, int y, uint8_t w);
 void set_Pixel(Pixel pixel);
 
 
-void draw_line(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1);
+void draw_line(int x0, int y0, int x1, int y1);
 void draw_rect(uint8_t x, uint8_t y, uint8_t width, uint8_t height, uint8_t r, uint8_t g, uint8_t b);
 void draw_Rect(Rect rect, uint8_t r, uint8_t g, uint8_t b);
 void Draw_Rect(Rect rect, Color color);
